+++ app/controllers/all/trusts/projects_controller.rb
@@ -1,8 +1,13 @@
-class All::Trust::ProjectsController < ApplicationController
+class All::Trusts::ProjectsController < ApplicationController
   after_action :verify_authorized
   rescue_from ActiveRecord::RecordNotFound, with: :not_found_error
 
-  def by_trust
+  def index
+    authorize Project, :index?
+    @pager, @trusts = pagy_array(ByTrustProjectFetcherService.new.call)
+  end
+
+  def show
     authorize Project, :index?
     ukprn = params[:trust_ukprn]
     @pager, @projects = pagy(Conversion::Project.by_trust_ukprn(ukprn).by_conversion_date)